requirements.txt
jinja2
pyyaml
openai
requests
